backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
google-generativeai
itsdangerous
httpx
